requirements.txt
# MRAF-Net Requirements
# Python 3.10+ required

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Medical Imaging
monai>=1.3.0
nibabel>=5.0.0
SimpleITK>=2.3.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Data Processing
pandas>=2.0.0
h5py>=3.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Progress and Logging
tqdm>=4.66.0
colorama>=0.4.6
rich>=13.5.0

# Utilities
einops>=0.7.0
timm>=0.9.0

# Optional: For Jupyter notebooks
jupyter>=1.0.0
ipywidgets>=8.1.0
